Subscribe / Unsubscribe mostly working
Email confirmation sending commented out for now,
but should work when it isn't (though we need to look
into alternative SMTP/sendgrid/postmark type API)

Subscribing gives you a sensible page when you
don't supply a redirect, and redirects if you do.

Link in confirmation mail works, but hasn't got a
redirect ability yet. Should do. Not sure how right
now. A field in the Maillist table presumably, pass
paramters to it.

Next up is doing that, and then we'll need to allow
actually sending email to the list.