Many updates
Messages are being sent.

Main todo things are SMTP settings and letting
the emails get sent at an accelerating rate.

I guess that should be per-message, but we have
a global "rate" in the display currently set to 1.

So is a send-rate set per-message?

If so it needs an average or soemting where it
currently have a 1