Inline formatting for buttons.
Email on view/sent more similar, all joined
into one function as it should have been.

Except the share buttons. Too much clutter for
the email-footer.

Inline formatting for the buttons coz gmail is
basically all of gmail and gmail is a bitch
that won't do style-sheets properly

React-counters moved to the bottom and restyled