Seems like a good bookmark. Though dunno what I've been doing.
Recently done is allowing a limit on the number
of emails that can be sent in a single day. Which
has left it all in a state where both sendjobs
and members can be in state=-2

IE, we need to do that tomorrow because we have
run out of our quota of email-sending for the day
(or, to you in particular (todo))

So the next job is to process those things
every new day.

What determines a new day?

Currently deep in the bowels of
the limit-checking function.

Which is not right at all.

So we need to move that,

And then add a "retry all the -2 things"
function in that new place which isn't
so obscure.

There's loads of other stuff here too.

Sorry.

Working alone and half drunk. Not really
keeping the history or logs as straight
or clean as they ought to be.