Max-sending per day limits.
Which also means we have to have an overflow queue
of sorts. Members and sendjobs go into state=-2 if
we try to send a message but hit the max-per-day limit.

Next day, we reset all the sendjobs with state=-2
and also send confirmation emails to all the people
with state=-2

Plus we give the messages sendrate a bump if we
add new emails to their send-queue.