Add a captcha
Sort of a captcha anyway, just looks for
saying "hello". A robot which was designed to
overcome this won't have any trouble at all
but I assume most of these spam subscriptions
are just general robots wondering around looking
for login forms to try and break.

So it might stop most of them.