Add a system to show warnings in Settings interface
Bug reporters never know what's going on, hopefully
telling them to reload the page and see if there is
a new latest-warning in the settings page will help
figure out what's going on.

Meanwhile, also fixed the packaging system so that
the vendor files which are needed are included. Should
have figured that out earlier instead of just trusting
some random who said not to include anything in the vendor
directory.