Fix some regression errors
I dunno what that $counter was even supposed to be,
but it was undefined.

I think getSuburl was maybe missing brackets? Dunno
what happened with that but  it should be fixed?

I also fixed errors to do with newest PHP not
liking nulls where it expected strings unless
you add a ? apparently?

I have no idea what I'm doing.

Ref:
https://github.com/revpriest/listman/issues/23